WebSockets/tests/submissions/Opera/interfaces/WebSocket/url/006.html

<!doctype html>
<title>WebSockets: defineProperty setter for url</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function() {
  assert_throws(new TypeError(), 
    function(){
      Object.defineProperty(WebSocket.prototype, 'readyState', {
        set: function() { return 'foo'; }
      });
    }
  );
});
</script>
